+++ Test/_fixtures/product_review_data.php
@@ -1,0 +1,42 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+return [
+    [
+        'id' => '4',
+        'articleID' => '198',
+        'name' => 'Pep Eroni',
+        'headline' => 'Hervorragend',
+        'comment' => 'bin sehr zufrieden...',
+        'points' => '5',
+        'datum' => '2012-08-22 10:29:07',
+        'active' => '1',
+        'email' => '',
+        'answer' => 'Danke',
+        'answer_date' => '2012-08-29 13:13:19',
+        'shop_id' => null,
+        'mainShopId' => '1',
+        '_locale' => 'de-DE',
+    ],
+
+    [
+        'id' => '14',
+        'articleID' => '145',
+        'name' => 'Das ist ein Name',
+        'headline' => 'Geiles Teil',
+        'comment' => 'fsfsdfsdfsdfsdfsdfsdf',
+        'points' => '5',
+        'datum' => '2020-07-28 12:12:02',
+        'active' => '1',
+        'email' => 'max@mustermann.de',
+        'answer' => '',
+        'answer_date' => null,
+        'shop_id' => '1',
+        'mainShopId' => '1',
+        '_locale' => 'de-DE',
+    ],
+];